Poker bot of 2003, bot written in MFC/C++
Perhaps HfStd.cpp:DumpScreenshot() and ImageAnalyzer::AnalyzeBitmap() are
among the most noteworthy. It's fast and should be general enough that it
works with most card games (possibly requiring minor tweaks, can't
remember).

It should also be easy to port to a generic Win32 app, nothing within MFC
is really used apart from setting up a debug window, which shows what
analyzed hands.

The HfPokerGameAnalyzer is written for Seven Card Stud, in which up to
four of each opponent's cards are visible. That class probably need some
updating if you want to use it for Hold'em or other card games.

Clicking 'call', 'raise', 'fold', 'check' and 'ante' are located in
PALZDLG.cpp.